Tidy up and move sources
