+++ src/ripple/ledger/ReadView.h
@@ -310,18 +310,6 @@ public:
     digest (key_type const& key) const = 0;
 };
 
-/** Run a functor on each SLE in a ReadView starting from the key start,
-    as long as the functor returns true.
- */
-template <class Functor>
-void forEachSLE(ReadView const& view, Functor func, uint256 const& start = {})
-{
-    for (auto k = view.succ(start); k; k = view.succ(*k))
-        if (auto sle = view.read(keylet::unchecked(*k)))
-            if (! func(*sle))
-                break;
-}
-
 } // ripple
 
 #include <ripple/ledger/detail/ReadViewFwdRange.ipp>
