+++ src/ripple/rpc/handlers/Tx.cpp
@@ -63,6 +63,26 @@ isValidated (RPC::Context& context, std::uint32_t seq, uint256 const& hash)
     return context.ledgerMaster.getHashBySeq (seq) == hash;
 }
 
+bool
+getMetaHex (Ledger const& ledger,
+    uint256 const& transID, std::string& hex)
+{
+    SHAMapTreeNode::TNType type;
+    auto const item =
+        ledger.txMap().peekItem (transID, type);
+
+    if (!item)
+        return false;
+
+    if (type != SHAMapTreeNode::tnTRANSACTION_MD)
+        return false;
+
+    SerialIter it (item->slice());
+    it.getVL (); // skip transaction
+    hex = strHex (makeSlice(it.getVL ()));
+    return true;
+}
+
 Json::Value doTx (RPC::Context& context)
 {
     if (!context.params.isMember (jss::transaction))
