+++ src/ripple/rpc/impl/Utilities.cpp
@@ -20,6 +20,8 @@
 #include <ripple/rpc/impl/Utilities.h>
 #include <ripple/json/json_value.h>
 #include <ripple/protocol/JsonFields.h>
+#include <boost/algorithm/string.hpp>
+#include <boost/format.hpp>
 
 namespace ripple {
 namespace RPC {
@@ -63,6 +65,33 @@ addPaymentDeliveredAmount (
     // Otherwise we report "unavailable" which cannot be parsed into a
     // sensible amount.
     meta[jss::delivered_amount] = Json::Value ("unavailable");
+}
+
+void
+injectSLE (Json::Value& jv,
+    SLE const& sle)
+{
+    jv = sle.getJson(0);
+    if (sle.getType() == ltACCOUNT_ROOT)
+    {
+        if (sle.isFieldPresent(sfEmailHash))
+        {
+            auto const& hash =
+                sle.getFieldH128(sfEmailHash);
+            Blob const b (hash.begin(), hash.end());
+            std::string md5 = strHex(makeSlice(b));
+            boost::to_lower(md5);
+            // VFALCO TODO Give a name and move this constant
+            //             to a more visible location. Also
+            //             shouldn't this be https?
+            jv[jss::urlgravatar] = str(boost::format(
+                "http://www.gravatar.com/avatar/%s") % md5);
+        }
+    }
+    else
+    {
+        jv[jss::Invalid] = true;
+    }
 }
 
 } // ripple
