+++ src/ripple/rpc/impl/Utilities.h
@@ -20,7 +20,7 @@
 #ifndef RIPPLE_RPC_IMPL_UTILITIES_H_INCLUDED
 #define RIPPLE_RPC_IMPL_UTILITIES_H_INCLUDED
 
-#include <ripple/app/ledger/TxMeta.h>
+#include <ripple/ledger/TxMeta.h>
 #include <ripple/app/tx/Transaction.h>
 
 namespace Json {
@@ -39,6 +39,18 @@ addPaymentDeliveredAmount (
     Context&,
     Transaction::pointer,
     TxMeta::pointer);
+
+/** Inject JSON describing ledger entry
+
+    Effects:
+        Adds the JSON description of `sle` to `jv`.
+
+        If `sle` holds an account root, also adds the
+        urlgravatar field JSON if sfEmailHash is present.
+*/
+void
+injectSLE (Json::Value& jv,
+    SLE const& sle);
 
 } // RPC
 } // ripple
