+++ src/ripple/unity/app_ledger.cpp
@@ -40,7 +40,6 @@
 #include <ripple/app/ledger/impl/LedgerTiming.cpp>
 #include <ripple/app/ledger/impl/OpenLedger.cpp>
 #include <ripple/app/ledger/impl/LedgerToJson.cpp>
-#include <ripple/app/ledger/impl/TxMeta.cpp>
 
 #include <ripple/app/ledger/tests/common_ledger.cpp>
 #include <ripple/app/ledger/tests/Ledger_test.cpp>
